feat: Add Docker build and publish workflow
This commit adds a new YAML file `docker_build.yml` that defines a workflow for building and publishing containers. The workflow is triggered on push events to the `main` branch. It sets up environment variables for the registry and image name, runs on Ubuntu latest, and has necessary permissions.

The steps in the workflow include checking out the repository, setting up Docker buildx, logging into the specified registry using GitHub actor credentials, and building/pushing the Docker image using Buildx. The image is built for both Linux AMD64 and ARM64 platforms.

The resulting Docker image is tagged as `ghcr.io/hra42/teams-background-generator:latest`.